Added dockerfile and docker compose

+++ Scrap_Autoria/AutoRia/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from sqlalchemy.orm import DeclarativeBase
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, BigInteger
 
 
 class Base(DeclarativeBase):
@@ -13,10 +13,10 @@ class UsedCar(Base):
     id = Column(Integer, primary_key=True)
     url = Column(String)
     title = Column(String)
-    price_usd = Column(Integer)
-    odometer = Column(Integer, default=0)
+    price_usd = Column(BigInteger)
+    odometer = Column(BigInteger, default=0)
     username = Column(String)
-    phone_number = Column(Integer)
+    phone_number = Column(BigInteger)
     image_url = Column(String, default="Not found")
     images_count = Column(Integer)
     car_number = Column(String, default="Not found")
